refactor: Move consumer code out of command namespace
- Move core functionality out of Command class, to module top level, so
  that testing and calling will be easier.
- Move code to new event_consumer module to mirror event_producer.
- Create a shim management-command module for the app, forwarding to the
  Command (now called ConsumeEventsCommand), since apps.py is now in a
  different place relative to the code.
