+++ edx_event_bus_kafka/management/commands/consume_events.py
@@ -1,0 +1,5 @@
+"""
+Makes ``consume_events`` management command available.
+"""
+
+from edx_event_bus_kafka.consumer.event_consumer import ConsumeEventsCommand as Command
